+++ CHANGELOG.md
@@ -52,6 +52,9 @@ The format is based on [Keep a Changelog].
   default is t) ([#261]).
 
 ### Enhancements
+* The default value of `selectrum-extend-current-candidate-highlight`
+  has been changed to `auto` which will automatically extend the
+  highlighting if the session uses annotations ([#427], [#430]).
 * The accessibility of the default value in file completions has been
   improved. When the default exists in the prompting directory it gets
   sorted first otherwise the default is included as a virtual
@@ -361,6 +364,8 @@ The format is based on [Keep a Changelog].
 [#414]: https://github.com/raxod502/selectrum/pull/414
 [#420]: https://github.com/raxod502/selectrum/issues/420
 [#421]: https://github.com/raxod502/selectrum/pull/421
+[#427]: https://github.com/raxod502/selectrum/issues/427
+[#430]: https://github.com/raxod502/selectrum/pull/430
 
 ## 3.0 (released 2020-10-20)
 ### Breaking changes